implementing limit to the quantity of stored logs
The current implementation does not care how many logs have been stored
in the buffer. This would allow applications to have an infinite amount
of logs stored in the buffer if they never become visible due to an
error log.
By implementing a ring buffer this would not be a problem, since it
would limit the amount of logs that can be stored at once. This limit is
done by removing the oldest stored log when a new log message is added.
Currently there is no way to change the total size of the buffer, which
is set to 20.

For more informations check issue #4.